--- src/components/atoms/paperBadge.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-
-const PaperBadge = ({ text, badgeClass }) => (
-  <div>
-    <span className={badgeClass}>{text}</span>
-  </div>
-)
-
-
-export default PaperBadge
